fixup! remove utils methods and duplicated methods
