+++ systemtest/src/main/java/io/strimzi/systemtest/utils/kafkaUtils/KafkaUtils.java
@@ -10,6 +10,8 @@ import io.strimzi.api.kafka.model.status.Condition;
 import io.strimzi.api.kafka.model.status.ListenerStatus;
 import io.strimzi.systemtest.Constants;
 import io.strimzi.systemtest.resources.crd.KafkaResource;
+import io.strimzi.systemtest.utils.kubeUtils.controllers.DeploymentUtils;
+import io.strimzi.systemtest.utils.kubeUtils.controllers.StatefulSetUtils;
 import io.strimzi.test.TestUtils;
 import io.strimzi.test.k8s.exceptions.KubeClusterException;
 import org.apache.logging.log4j.LogManager;
@@ -18,9 +20,12 @@ import org.apache.logging.log4j.Logger;
 import java.nio.charset.Charset;
 import java.util.Base64;
 import java.util.List;
+import java.util.Map;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
+import static io.strimzi.api.kafka.model.KafkaResources.kafkaStatefulSetName;
+import static io.strimzi.api.kafka.model.KafkaResources.zookeeperStatefulSetName;
 import static io.strimzi.test.TestUtils.indent;
 import static io.strimzi.test.TestUtils.waitFor;
 import static io.strimzi.test.k8s.KubeClusterResource.cmdKubeClient;
@@ -118,5 +123,45 @@ public class KafkaUtils {
         secretCerts = new String(decodedBytes, Charset.defaultCharset());
 
         return secretCerts;
+    }
+
+    @SuppressWarnings("unchecked")
+    public static void waitForClusterStability(String clusterName) {
+        LOGGER.info("Waiting for cluster stability");
+        Map<String, String>[] zkPods = new Map[1];
+        Map<String, String>[] kafkaPods = new Map[1];
+        Map<String, String>[] eoPods = new Map[1];
+        int[] count = {0};
+        zkPods[0] = StatefulSetUtils.ssSnapshot(zookeeperStatefulSetName(clusterName));
+        kafkaPods[0] = StatefulSetUtils.ssSnapshot(kafkaStatefulSetName(clusterName));
+        eoPods[0] = DeploymentUtils.depSnapshot(KafkaResources.entityOperatorDeploymentName(clusterName));
+        TestUtils.waitFor("Cluster stable and ready", Constants.GLOBAL_POLL_INTERVAL, Constants.TIMEOUT_FOR_CLUSTER_STABLE, () -> {
+            Map<String, String> zkSnapshot = StatefulSetUtils.ssSnapshot(zookeeperStatefulSetName(clusterName));
+            Map<String, String> kafkaSnaptop = StatefulSetUtils.ssSnapshot(kafkaStatefulSetName(clusterName));
+            Map<String, String> eoSnapshot = DeploymentUtils.depSnapshot(KafkaResources.entityOperatorDeploymentName(clusterName));
+            boolean zkSameAsLast = zkSnapshot.equals(zkPods[0]);
+            boolean kafkaSameAsLast = kafkaSnaptop.equals(kafkaPods[0]);
+            boolean eoSameAsLast = eoSnapshot.equals(eoPods[0]);
+            if (!zkSameAsLast) {
+                LOGGER.info("ZK Cluster not stable");
+            }
+            if (!kafkaSameAsLast) {
+                LOGGER.info("Kafka Cluster not stable");
+            }
+            if (!eoSameAsLast) {
+                LOGGER.info("EO not stable");
+            }
+            if (zkSameAsLast
+                    && kafkaSameAsLast
+                    && eoSameAsLast) {
+                int c = count[0]++;
+                LOGGER.info("All stable for {} polls", c);
+                return c > 60;
+            }
+            zkPods[0] = zkSnapshot;
+            kafkaPods[0] = kafkaSnaptop;
+            count[0] = 0;
+            return false;
+        });
     }
 }
