--- systemtest/src/main/java/io/strimzi/systemtest/utils/kubeUtils/objects/ClusterUtils.java
@@ -1,63 +1,0 @@
-/*
- * Copyright Strimzi authors.
- * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
- */
-package io.strimzi.systemtest.utils.kubeUtils.objects;
-
-import io.strimzi.api.kafka.model.KafkaResources;
-import io.strimzi.systemtest.Constants;
-import io.strimzi.systemtest.utils.kubeUtils.controllers.DeploymentUtils;
-import io.strimzi.systemtest.utils.kubeUtils.controllers.StatefulSetUtils;
-import io.strimzi.test.TestUtils;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-import java.util.Map;
-import java.util.concurrent.atomic.AtomicInteger;
-
-import static io.strimzi.api.kafka.model.KafkaResources.kafkaStatefulSetName;
-import static io.strimzi.api.kafka.model.KafkaResources.zookeeperStatefulSetName;
-
-public class ClusterUtils {
-    private static final Logger LOGGER = LogManager.getLogger(ClusterUtils.class);
-
-    @SuppressWarnings("unchecked")
-    public static void waitForClusterStability(String clusterName) {
-        LOGGER.info("Waiting for cluster stability");
-        Map<String, String>[] zkPods = new Map[1];
-        Map<String, String>[] kafkaPods = new Map[1];
-        Map<String, String>[] eoPods = new Map[1];
-        AtomicInteger count = new AtomicInteger();
-        zkPods[0] = StatefulSetUtils.ssSnapshot(zookeeperStatefulSetName(clusterName));
-        kafkaPods[0] = StatefulSetUtils.ssSnapshot(kafkaStatefulSetName(clusterName));
-        eoPods[0] = DeploymentUtils.depSnapshot(KafkaResources.entityOperatorDeploymentName(clusterName));
-        TestUtils.waitFor("Cluster stable and ready", Constants.GLOBAL_POLL_INTERVAL, Constants.TIMEOUT_FOR_CLUSTER_STABLE, () -> {
-            Map<String, String> zkSnapshot = StatefulSetUtils.ssSnapshot(zookeeperStatefulSetName(clusterName));
-            Map<String, String> kafkaSnaptop = StatefulSetUtils.ssSnapshot(kafkaStatefulSetName(clusterName));
-            Map<String, String> eoSnapshot = DeploymentUtils.depSnapshot(KafkaResources.entityOperatorDeploymentName(clusterName));
-            boolean zkSameAsLast = zkSnapshot.equals(zkPods[0]);
-            boolean kafkaSameAsLast = kafkaSnaptop.equals(kafkaPods[0]);
-            boolean eoSameAsLast = eoSnapshot.equals(eoPods[0]);
-            if (!zkSameAsLast) {
-                LOGGER.info("ZK Cluster not stable");
-            }
-            if (!kafkaSameAsLast) {
-                LOGGER.info("Kafka Cluster not stable");
-            }
-            if (!eoSameAsLast) {
-                LOGGER.info("EO not stable");
-            }
-            if (zkSameAsLast
-                    && kafkaSameAsLast
-                    && eoSameAsLast) {
-                int c = count.getAndIncrement();
-                LOGGER.info("All stable for {} polls", c);
-                return c > 60;
-            }
-            zkPods[0] = zkSnapshot;
-            kafkaPods[0] = kafkaSnaptop;
-            count.set(0);
-            return false;
-        });
-    }
-}
